Move stringref id and refcount to unsigned int. 16 bits was not enough to pass value_stresstest2.

+++ include/grace/statstring.h
@@ -32,8 +32,8 @@ struct stringref
 	stringref					*parent; ///< Link to parent node in the hash tree.
 	stringref					*lower, *higher; ///< Link to siblings in the hash tree.
 	volatile unsigned int		 key; ///< This string's hash key.
-	volatile unsigned short		 id; ///< This stringref's unique id.
-	volatile unsigned short		 refcnt; ///< Reference count.
+	volatile unsigned int		 id; ///< This stringref's unique id.
+	volatile unsigned int		 refcnt; ///< Reference count.
 	string						 str; ///< Actual string data.
 };
 
@@ -106,9 +106,9 @@ public:
 							 /// gets a unique id. This is implemented
 							 /// using a simple sequence counter.
 							 /// \return new id.
-	unsigned short			 newid (void)
+	unsigned int			 newid (void)
 							 {
-							 	unsigned short res;
+							 	unsigned int res;
 							 	res = sequence++;
 							 	
 							 	return res;
@@ -117,7 +117,7 @@ public:
 protected:
 	stringref				*root; ///< Root node of the database.
 	stringref				*nukeroot; ///< Root node of deleted entries
-	unsigned short			 sequence; ///< Current sequence number.
+	unsigned int			 sequence; ///< Current sequence number.
 	lock<int>				 treelock; ///< Lock for the database.
 	
 							 /// Remove a reference from the tree.
